test: fix try/finally anti-pattern and unguarded cleanup in tests
Move WhenPathExistsOnDisk_FormatsNormally into FormatWithBasePath class
to use its IDisposable fixture instead of try/finally. Wrap act+assert
in WhenAbsoluteFlag_OutputsAbsolutePath with try/finally so Directory.Delete
runs even when assertions fail.
